added max cap for annual base on user target amount

+++ src/main/java/com/example/MobilePaluwagan/DTOs/Response/SavingsSummaryResponse.java
@@ -5,14 +5,15 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class SavingsSummaryResponse {
-    private double totalSavingsBalance;
-    private double annualMoney;
+    private BigDecimal totalSavingsBalance;
+    private BigDecimal annualMoney;
     private String savingsId;
     private List<SavingsDepositHistory> depositHistoryList;
 }
